v2.8.0

+++ src/JordiAragon.Cinema.Reservation.Presentation.HttpRestfulApi.Contracts.V2/Auditorium/Responses/SeatResponse.cs
@@ -2,5 +2,5 @@
 {
     using System;
 
-    public sealed record class SeatResponse(Guid Id, int Row, int SeatNumber);
+    public sealed record class SeatResponse(Guid Id, ushort Row, ushort SeatNumber);
 }
